requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2

langchain==0.3.0
langchain-openai==0.2.0
langchain-community==0.3.0

faiss-cpu==1.8.0.post1
tiktoken==0.7.0